requirements.txt
pyTelegramBotAPI~=4.17.0
python-dotenv~=1.0.1